+++ .env
@@ -1,0 +1,4 @@
+POSTGRES_USER=radio
+POSTGRES_PASSWORD=sondedb
+
+PYTHONUNBUFFERED=true
