+++ nginx/web.conf
@@ -1,0 +1,15 @@
+server {
+    listen 80;
+    server_name _;
+
+    location / {
+        proxy_pass http://web:5000;
+        proxy_redirect off;
+
+        proxy_set_header   Host                 $host;
+        proxy_set_header   X-Real-IP            $remote_addr;
+        proxy_set_header   X-Forwarded-For      $proxy_add_x_forwarded_for;
+        proxy_set_header   X-Forwarded-Proto    $scheme;
+
+    }
+}
