+++ radiosonde/wsgi.py
@@ -1,0 +1,4 @@
+from .app import app
+
+if __name__ == "__main__":
+    app.run()
